# docs/technical/03_api_specification.generated.yaml
components:
  schemas:
...
    DonorSponsorshipCreate:
      properties:
        currency:
          maxLength: 3
          minLength: 3
          title: Currency
          type: string
        frequency:
          default: monthly
          enum:
          - monthly
          - quarterly
          - semi_annual
          - annual
          - one_time
          title: Frequency
          type: string
        monthly_amount:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,8}|(?=[\d.]{1,11}0*$)\d{0,8}\.\d{0,2}0*$)
            type: string
          title: Monthly Amount
        orphan_code:
          maxLength: 20
          minLength: 1
          title: Orphan Code
          type: string
      required:
      - orphan_code
      - monthly_amount
      - currency
      title: DonorSponsorshipCreate
      type: object
    DonorSponsorshipMini:
      description: Lean projection — what the donor portal needs to render.
      properties:
        code:
          title: Code
          type: string
        currency:
          title: Currency
          type: string
        id:
          format: uuid
          title: Id
          type: string
        monthly_amount:
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Monthly Amount
          type: string
        orphan_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Orphan Code
        orphan_first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Orphan First Name
        start_date:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Start Date
        status:
          title: Status
          type: string
        total_paid:
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Paid
          type: string
      required:
      - id
      - code
      - monthly_amount
      - currency
      - status
      - start_date
      - total_paid
      - orphan_code
      - orphan_first_name
      title: DonorSponsorshipMini
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Page_SponsorshipRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SponsorshipRead'
          title: Items
          type: array
        limit:
          title: Limit
          type: integer
        offset:
          title: Offset
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: Page[SponsorshipRead]
      type: object
...
    SponsorshipCancel:
      properties:
        reason:
          anyOf:
          - maxLength: 1000
            type: string
          - type: 'null'
          title: Reason
      title: SponsorshipCancel
      type: object
    SponsorshipCreate:
      properties:
        currency:
          maxLength: 3
          minLength: 3
          title: Currency
          type: string
        donor_id:
          format: uuid
          title: Donor Id
          type: string
        end_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: End Date
        monthly_amount:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,8}|(?=[\d.]{1,11}0*$)\d{0,8}\.\d{0,2}0*$)
            type: string
          title: Monthly Amount
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        orphan_id:
          format: uuid
          title: Orphan Id
          type: string
        payment_frequency:
          default: monthly
          enum:
          - monthly
          - quarterly
          - semi_annual
          - annual
          - one_time
          title: Payment Frequency
          type: string
        payment_method:
          anyOf:
          - maxLength: 30
            type: string
          - type: 'null'
          title: Payment Method
        start_date:
          format: date
          title: Start Date
          type: string
      required:
      - donor_id
      - orphan_id
      - monthly_amount
      - currency
      - start_date
      title: SponsorshipCreate
      type: object
    SponsorshipRead:
      properties:
        cancelled_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Cancelled At
        code:
          title: Code
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        currency:
          maxLength: 3
          minLength: 3
          title: Currency
          type: string
        donor_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Donor Code
        donor_id:
          format: uuid
          title: Donor Id
          type: string
        donor_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Donor Name
        end_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: End Date
        id:
          format: uuid
          title: Id
          type: string
        monthly_amount:
          pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,8}|(?=[\d.]{1,11}0*$)\d{0,8}\.\d{0,2}0*$)
          title: Monthly Amount
          type: string
        next_payment_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Next Payment Date
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        organization_id:
          format: uuid
          title: Organization Id
          type: string
        orphan_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Orphan Code
        orphan_id:
          format: uuid
          title: Orphan Id
          type: string
        orphan_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Orphan Name
        payment_frequency:
          default: monthly
          enum:
          - monthly
          - quarterly
          - semi_annual
          - annual
          - one_time
          title: Payment Frequency
          type: string
        payment_method:
          anyOf:
          - maxLength: 30
            type: string
          - type: 'null'
          title: Payment Method
        payments_count:
          title: Payments Count
          type: integer
        start_date:
          format: date
          title: Start Date
          type: string
        status:
          enum:
          - pending
          - active
          - paused
          - overdue
          - cancelled
          - completed
          - transferred
          title: Status
          type: string
        total_paid:
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Total Paid
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - donor_id
      - orphan_id
      - monthly_amount
      - currency
      - start_date
      - id
      - code
      - organization_id
      - status
      - total_paid
      - payments_count
      - next_payment_date
      - cancelled_at
      - created_at
      - updated_at
      title: SponsorshipRead
      type: object
    SponsorshipUpdate:
      description: 'Partial update. Status changes go through dedicated /pause and

        /resume endpoints — leaving the cancel/complete transitions on

        their own routes so each one can keep its specific guard logic.'
      properties:
        end_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: End Date
        monthly_amount:
          anyOf:
          - exclusiveMinimum: 0.0
            type: number
          - pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,8}|(?=[\d.]{1,11}0*$)\d{0,8}\.\d{0,2}0*$)
            type: string
          - type: 'null'
          title: Monthly Amount
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        payment_frequency:
          anyOf:
          - enum:
            - monthly
            - quarterly
            - semi_annual
            - annual
            - one_time
            type: string
          - type: 'null'
          title: Payment Frequency
        payment_method:
          anyOf:
          - maxLength: 30
            type: string
          - type: 'null'
          title: Payment Method
      title: SponsorshipUpdate
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  title: Rufaqaa API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/donor/me/sponsorships:
    get:
      operationId: list_my_sponsorships_api_v1_donor_me_sponsorships_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DonorSponsorshipMini'
                title: Response List My Sponsorships Api V1 Donor Me Sponsorships Get
                type: array
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: List My Sponsorships
      tags:
      - donor-self
    post:
      description: 'Donor-driven sponsorship creation. The orphan is looked up by

        its public code (not internal UUID) so the donor never sees it.

        Only orphans that are publicly browseable can be sponsored.'
      operationId: create_my_sponsorship_api_v1_donor_me_sponsorships_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DonorSponsorshipCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonorSponsorshipMini'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create My Sponsorship
      tags:
      - donor-self
...
  /api/v1/me/sponsorships:
    get:
      operationId: my_sponsorships_api_v1_me_sponsorships_get
      parameters:
      - in: query
        name: donor_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Donor Id
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_SponsorshipRead_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: My Sponsorships
      tags:
      - donor-portal
...
  /api/v1/sponsorships:
    get:
      operationId: list_sponsorships_api_v1_sponsorships_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: donor_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Donor Id
      - in: query
        name: orphan_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Orphan Id
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_SponsorshipRead_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: List Sponsorships
      tags:
      - sponsorships
    post:
      operationId: create_sponsorship_api_v1_sponsorships_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SponsorshipCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorshipRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Sponsorship
      tags:
      - sponsorships
  /api/v1/sponsorships/export.csv:
    get:
      description: 'Stream sponsorships matching the same filters as the list endpoint,

        as CSV. Registered before /{sponsorship_id} so FastAPI doesn''t try

        to parse ''export.csv'' as a UUID. Capped at 10 000 rows.'
      operationId: export_sponsorships_csv_api_v1_sponsorships_export_csv_get
      parameters:
      - in: query
        name: donor_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Donor Id
      - in: query
        name: orphan_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Orphan Id
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Export Sponsorships Csv
      tags:
      - sponsorships
  /api/v1/sponsorships/{sponsorship_id}:
    get:
      operationId: get_sponsorship_api_v1_sponsorships__sponsorship_id__get
      parameters:
      - in: path
        name: sponsorship_id
        required: true
        schema:
          format: uuid
          title: Sponsorship Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorshipRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Sponsorship
      tags:
      - sponsorships
    patch:
      description: 'Edit non-status fields on an active sponsorship.


        Cancelled / completed sponsorships are immutable so the historical

        record stays clean.'
      operationId: update_sponsorship_api_v1_sponsorships__sponsorship_id__patch
      parameters:
      - in: path
        name: sponsorship_id
        required: true
        schema:
          format: uuid
          title: Sponsorship Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SponsorshipUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorshipRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Sponsorship
      tags:
      - sponsorships
  /api/v1/sponsorships/{sponsorship_id}/cancel:
    post:
      operationId: cancel_sponsorship_api_v1_sponsorships__sponsorship_id__cancel_post
      parameters:
      - in: path
        name: sponsorship_id
        required: true
        schema:
          format: uuid
          title: Sponsorship Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SponsorshipCancel'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorshipRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Cancel Sponsorship
      tags:
      - sponsorships
  /api/v1/sponsorships/{sponsorship_id}/pause:
    post:
      operationId: pause_sponsorship_api_v1_sponsorships__sponsorship_id__pause_post
      parameters:
      - in: path
        name: sponsorship_id
        required: true
        schema:
          format: uuid
          title: Sponsorship Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorshipRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Pause Sponsorship
      tags:
      - sponsorships
  /api/v1/sponsorships/{sponsorship_id}/resume:
    post:
      operationId: resume_sponsorship_api_v1_sponsorships__sponsorship_id__resume_post
      parameters:
      - in: path
        name: sponsorship_id
        required: true
        schema:
          format: uuid
          title: Sponsorship Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SponsorshipRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Resume Sponsorship
      tags:
      - sponsorships
...